Move HasFault to vim25/types
